refactor: Add observedGeneration status field and cleanup cache locks
This introduces an `observedGeneration` field to the status so we can properly events that are noop.
For example, when we update the checksum and lastCheck fields of the EnvWatcher resource, this triggers
an event, however these changes do not impact the content of the actual resource ( no generation change ).
By tracking the last Generation weve worked on, we can effectively disregard these events.

This also cleans up the context and locking around it. Previously we were handling the caching at the
top level Reconcile func, but this was unnecessary and caused flakes downstream depending on timing.
This moves the context cancelFunc to be in the watcher scope so we can be more deliberate with our
interactions.

Signed-off-by: Brad Beam <brad.beam@b-rad.info>